Extract GC-agnostic components of SWTGraphics into new class
This introduces a new `NativeGraphics` class, which inherits all code of
the `SWTGraphics` class, which doesn't directly depend on a GC object.
Doing so drastically reduces the amount of code that needs to be written
when introducing an `AWTGraphics` class, while at the same time
minimizing the chance of deviations between both toolkits.

Transformation requires a special handling because of how it interacts
with the clipping. The GC-specifc part of the transformation has been
extracted into abstract "performX" methods, while the GC-agnostic part
remains in the base class.

Contributes to https://github.com/eclipse-gef/gef-classic/pull/952
